removed incorrect dependencies
