+++ .gitignore
@@ -1,0 +1,2 @@
+*.pyc
+__pycache__/
