--- pymedimage/__init__.py
@@ -1,15 +1,0 @@
-import os
-import sys
-# add root to PATH
-sys.path.insert(0, os.path.dirname(os.path.realpath(__file__)))
-
-# Set default logging handler to avoid "No handler found" warnings.
-import logging
-try:  # Python 2.7+
-    from logging import NullHandler
-except ImportError:
-    class NullHandler(logging.Handler):
-        def emit(self, record):
-            pass
-
-logging.getLogger(__name__).addHandler(NullHandler())
