--- pymedimage/local_features_stats/glcm.h
@@ -1,1 +1,0 @@
-
