--- pymedimage/showvolume.py
@@ -1,44 +1,0 @@
-""" 
-Visualization of 3d numpy array for debugging and visualization
-Downloaded from the internet 
-https://www.datacamp.com/community/tutorials/matplotlib-3d-volumetric-data#gs.v0bGR6M
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-
-def multi_slice_viewer(volume):
-    
-    remove_keymap_conflicts({'j', 'k'})
-    fig, ax = plt.subplots()
-    ax.volume = volume
-    ax.index = volume.shape[0] // 2
-    ax.imshow(volume[ax.index])
-    fig.canvas.mpl_connect('key_press_event', process_key)
-    plt.show()
-
-def process_key(event):
-    fig = event.canvas.figure
-    ax = fig.axes[0]
-    if event.key == 'j':
-        previous_slice(ax)
-    elif event.key == 'k':
-        next_slice(ax)
-    fig.canvas.draw()
-
-def previous_slice(ax):
-    volume = ax.volume
-    ax.index = (ax.index - 1) % volume.shape[0]  # wrap around using %
-    ax.images[0].set_array(volume[ax.index])
-
-def next_slice(ax):
-    volume = ax.volume
-    ax.index = (ax.index + 1) % volume.shape[0]
-    ax.images[0].set_array(volume[ax.index])
-
-def remove_keymap_conflicts(new_keys_set):
-    for prop in plt.rcParams:
-        if prop.startswith('keymap.'):
-            keys = plt.rcParams[prop]
-            remove_list = set(keys) & new_keys_set
-            for key in remove_list:
-                keys.remove(key)
